feat: optimize popup.html loading time
